+++ tests/test_bufremove.lua
@@ -107,22 +107,42 @@ local validate_unshow_with_buf_id = function(fun_name, layout)
   eq(win_get_buf(layout['win_right']), layout['buf_right'])
 end
 
+local mock_confirm = function(user_choice)
+  local lua_cmd = string.format('vim.fn.confirm = function(...) _G.confirm_args = { ... }; return %d end', user_choice)
+  child.lua(lua_cmd)
+end
+
 local validate_force_argument = function(fun_name, layout)
   child.api.nvim_buf_set_lines(layout['buf'], 0, -1, true, { 'aaa' })
   -- Avoid hit-enter prompt due to long message
   child.o.cmdheight = 10
 
+  -- Should ask for confirmation for modified buffer and no `force`
+  mock_confirm(1)
+  eq(child.bo.modified, true)
   local output = child.lua_get(('MiniBufremove.%s()'):format(fun_name))
   eq(output, false)
+
+  local prompt =
+    string.format('Buffer %d has unsaved changes. Do you want to force %s?', child.api.nvim_get_current_buf(), fun_name)
+  -- - Default choice should be 'No'
+  eq(child.lua_get('_G.confirm_args'), { prompt, '&No\n&Yes', 1, 'Question' })
+
   eq(win_get_buf(layout['win_left']), layout['buf'])
   eq(win_get_buf(layout['win_right']), layout['buf'])
 
-  local last_message = child.cmd_capture('1message')
-  expect.match(last_message, 'Buffer ' .. layout['buf'] .. ' has unsaved changes%..*Use.*force')
-
-  output = child.lua_get(('MiniBufremove.%s(nil, true)'):format(fun_name))
+  -- Should properly remove modified buffer after confirmation
+  mock_confirm(2)
+  eq(child.bo.modified, true)
+  output = child.lua_get(('MiniBufremove.%s()'):format(fun_name))
   eq(output, true)
   eq(win_get_buf(layout['win_left']), layout['buf_left'])
+  eq(win_get_buf(layout['win_right']), layout['buf_right'])
+
+  -- Should properly remove buffer with `force`
+  output = child.lua_get(('MiniBufremove.%s(nil, true)'):format(fun_name))
+  eq(output, true)
+  eq(win_get_buf(layout['win_left']), layout['buf_right'])
   eq(win_get_buf(layout['win_right']), layout['buf_right'])
 end